pyproject.toml
[project]
name = "zmq-req-rep-demo"
version = "0.1.0"
description = "REQ/REP server/client demo using pyzmq asyncio"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyzmq>=26.0.0",
]

[tool.uv]
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = []
